Handler split
Split the handler from the main typer class, and performed some assorted
small changes to make this a little lighter on the memory usage, just in
case. Additionally, added a garbage collector call right before the
setup of the activate handler in an attempt to get rid of any previous
versions of activate handler that may exist, to help decrease memory
usage. As stated in the Todo comments, next features to work on are
fixing the tab counts (they don't work with the current youtube layout)
and looking into having this keep the computer awake, so I don't have to
go open my mousejiggler myself.