Tab Changes
Currently, the tab counts only work for the initial video's details, (and that's discounting the video format popup sometimes not being there) and the second video. Afterwards, it types all the details into the tags. However, I'm out of videos to test this on, so I'll fix this proper when (ironically) I'm on a slower internet connection and have more time to work.