Feat/improvedlogging (#3)
* Initial commit

Added a file-based logging system, such that I can know what was happening (roughly) when the user crashed the program. Will be fixing the misaligned tabs in this branch, once I have videos to upload, to test with

* Tab Changes

Currently, the tab counts only work for the initial video's details, (and that's discounting the video format popup sometimes not being there) and the second video. Afterwards, it types all the details into the tags. However, I'm out of videos to test this on, so I'll fix this proper when (ironically) I'm on a slower internet connection and have more time to work.

* Log File Rename

Made the log file named properly. Can't believe I forgot to merge this into dev before I did further tinkering.