Functionality
After the program is finished typing things, access the command line, and give the appropriate set of commands based on system setup to set the sleep settings to not sleep at all. The command works in windows, but I've got no way to test this on mac or unix, so that's something to keep in mind.